+++ examples/object_geos_parts.rs
@@ -14,58 +14,58 @@ fn main() -> Result<()> {
 
     for info in node.get_objects_info()? {
         let obj_node = info.to_node()?;
-        if let Some(geometry) = obj_node.geometry()? {
-            for part_info in geometry.partitions()? {
-                println!(
-                    "Object: {}, Display: {}, Partition: {}",
-                    obj_node.path()?,
-                    geometry.node.path_relative(Some(obj_node.handle))?,
-                    part_info.part_id()
-                );
-                let attrib_names =
-                    geometry.get_attribute_names(AttributeOwner::Point, Some(&part_info))?;
-                println!(
-                    "{}",
-                    &attrib_names.iter_str().collect::<Vec<&str>>().join("\n")
-                );
-                println!("Point Positions: ");
-                let attrib = geometry
-                    .get_attribute(part_info.part_id(), AttributeOwner::Point, "P")?
-                    .unwrap();
+        let Some(geometry) = obj_node.geometry()? else {
+            continue;
+        };
+        for part_info in geometry.partitions()? {
+            println!(
+                "Object: {}, Display: {}, Partition: {}",
+                obj_node.path()?,
+                geometry.node.path_relative(Some(obj_node.handle))?,
+                part_info.part_id()
+            );
+            let attrib_names =
+                geometry.get_attribute_names(AttributeOwner::Point, Some(&part_info))?;
+            println!(
+                "{}",
+                &attrib_names.iter_str().collect::<Vec<&str>>().join("\n")
+            );
+            println!("Point Positions: ");
+            let attrib = geometry
+                .get_attribute(part_info.part_id(), AttributeOwner::Point, "P")?
+                .unwrap();
 
-                let attrib = attrib.downcast::<NumericAttr<f32>>().unwrap();
-                let positions = attrib.get(part_info.part_id())?;
-                for p in 0..attrib.info().count() {
-                    let idx = (p * attrib.info().tuple_size()) as usize;
-                    println!("{:?}", &positions[idx..idx + 3])
-                }
+            let attrib = attrib.downcast::<NumericAttr<f32>>().unwrap();
+            let positions = attrib.get(part_info.part_id())?;
+            for p in 0..attrib.info().count() {
+                let idx = (p * attrib.info().tuple_size()) as usize;
+                println!("{:?}", &positions[idx..idx + 3])
+            }
 
-                println!("Number of Faces: {}", part_info.face_count());
-                let faces = geometry.get_face_counts(Some(&part_info))?;
-                if part_info.part_type() != PartType::Curve {
-                    for face in faces.iter() {
-                        print!("{}, ", face);
-                    }
-                    println!();
+            println!("Number of Faces: {}", part_info.face_count());
+            let faces = geometry.get_face_counts(Some(&part_info))?;
+            if part_info.part_type() != PartType::Curve {
+                for face in faces.iter() {
+                    print!("{}, ", face);
                 }
-                let vertices = geometry.vertex_list(Some(&part_info))?;
-                println!("Vertex Indices Into Points Array");
-                let mut curr_idx = 0;
-                assert!(curr_idx < vertices.len());
-                for (face, count) in faces.iter().enumerate() {
-                    for _ in 0..(*count as usize) {
-                        println!(
-                            "Vertex: {0}, \
-                                  belonging to face: {1}, \
-                                  index: {2} of point array ",
-                            curr_idx, face, vertices[curr_idx]
-                        );
-                        curr_idx += 1;
-                    }
+                println!();
+            }
+            let vertices = geometry.vertex_list(Some(&part_info))?;
+            println!("Vertex Indices Into Points Array");
+            let mut curr_idx = 0;
+            assert!(curr_idx < vertices.len());
+            for (face, count) in faces.iter().enumerate() {
+                for _ in 0..(*count as usize) {
+                    println!(
+                        "Vertex: {0}, \
+                              belonging to face: {1}, \
+                              index: {2} of point array ",
+                        curr_idx, face, vertices[curr_idx]
+                    );
+                    curr_idx += 1;
                 }
             }
         }
     }
-
     Ok(())
 }